Keep the case in arrow schema.
