+++ src/query/storages/parquet/src/parquet_reader/meta.rs
@@ -17,9 +17,6 @@ use std::fs::File;
 
 use common_arrow::arrow::array::UInt64Array;
 use common_arrow::arrow::buffer::Buffer;
-use common_arrow::arrow::datatypes::DataType as ArrowDataType;
-use common_arrow::arrow::datatypes::Field as ArrowField;
-use common_arrow::arrow::datatypes::Schema as ArrowSchema;
 use common_arrow::arrow::io::parquet::read as pread;
 use common_arrow::parquet::metadata::FileMetaData;
 use common_arrow::parquet::metadata::RowGroupMetaData;
@@ -34,23 +31,6 @@ use storages_common_table_meta::meta::StatisticsOfColumns;
 
 use crate::ParquetReader;
 
-fn lower_field_name(field: &mut ArrowField) {
-    field.name = field.name.to_lowercase();
-    match &mut field.data_type {
-        ArrowDataType::List(f)
-        | ArrowDataType::LargeList(f)
-        | ArrowDataType::FixedSizeList(f, _) => {
-            lower_field_name(f.as_mut());
-        }
-        ArrowDataType::Struct(ref mut fields) => {
-            for f in fields {
-                lower_field_name(f);
-            }
-        }
-        _ => {}
-    }
-}
-
 impl ParquetReader {
     pub fn read_meta(location: &str) -> Result<FileMetaData> {
         let mut file = File::open(location).map_err(|e| {
@@ -62,15 +42,6 @@ impl ParquetReader {
                 location, e
             ))
         })
-    }
-
-    #[inline]
-    pub fn infer_schema(meta: &FileMetaData) -> Result<ArrowSchema> {
-        let mut arrow_schema = pread::infer_schema(meta)?;
-        arrow_schema.fields.iter_mut().for_each(|f| {
-            lower_field_name(f);
-        });
-        Ok(arrow_schema)
     }
 
     /// Collect statistics of a batch of row groups of the specified columns.
