+++ tests/suites/1_stateful/06_read_parquet/06_0001_complex.sh
